feat: Phase 1 코드 정리 작업 완료 (v4.2.7)
