+++ DEVELOPMENT_ROADMAP.md
@@ -7,7 +7,7 @@
 
 ## 📌 현재 프로젝트 상태
 
-### 완료된 작업 (v4.2.5 기준)
+### 완료된 작업 (v4.2.7 기준)
 - ✅ 주요 버그 수정 완료 (CLI help, URL 수집 제한, 날짜별 병합)
 - ✅ 문서 구조 정합성 확보
 - ✅ 403 오류 대응 시스템 구축
@@ -19,7 +19,7 @@
 1. **코드 구조** 
    - ~~테스트 파일들이 루트 디렉토리에 산재~~ → 일부 정리됨
    - ~~설정 파일 3개가 혼재~~ → unified_config.json 하나로 정리됨
-   - ~~print와 logger 혼용~~ → 대부분 해결됨
+   - ~~print와 logger 혼용~~ → 완전히 해결됨
 
 2. **미완성 기능**
    - ~~테스트 커버리지 부재~~ → 기본 테스트 환경 구축 완료
@@ -40,16 +40,16 @@
 - ✅ 미사용 import 제거 (전체 프로젝트)
 - ✅ PEP 8 스타일 적용 (newline, 간격, 포맷팅)
 - ✅ typing 모듈 정렬 및 최적화
-- ✅ 문서 버전 업데이트 (v4.2.5)
+- ✅ 문서 버전 업데이트 (v4.2.7)
 
-#### 🔄 1.2 다음 단계: 파일 구조 정리
+#### ✅ 1.2 파일 구조 정리 완료
 ```
 작업 내용:
-- 루트의 test_*.py 파일들을 tests/ 폴더로 이동 또는 삭제
-- 임시 파일 정리  
+- 루트의 test_*.py 파일들을 tests/ 폴더로 이동
+- 임시 파일 정리
 - 개발용 스크립트 분리
 
-우선순위: 중간 (Phase 2에서 진행)
+우선순위: 완료
 ```
 
 ### 우선순위 1: Phase 2 - 기능 완성 및 구조 개선
