+++ README.md
@@ -1,4 +1,4 @@
-# 네이버 뉴스 크롤러 v4.2.5
+# 네이버 뉴스 크롤러 v4.2.7
 
 네이버 뉴스 검색 결과를 수집하고 본문을 추출하는 Python 기반 크롤러입니다.
 
@@ -16,7 +16,7 @@
 - 🛡️ 403 오류 대응 및 세션 풀 관리 (v4.2)
 - ⚙️ 통합 설정 파일 시스템 (v4.2)
 - 🔄 개선된 날짜별 수집 및 병합 기능 (v4.2)
-- 🧹 코드 품질 개선 및 표준화 (v4.2.5)
+- 🧹 코드 품질 개선 및 표준화 (v4.2.7)
 
 ## 빠른 시작
