+++ docs/PROGRESS.md
@@ -3,7 +3,7 @@
 ## 프로젝트 개요
 - **목적**: 네이버 뉴스 검색 엔진을 통한 체계적인 뉴스 수집 및 본문 추출
 - **핵심 기능**: 검색어×언론사×날짜 매트릭스 방식의 체계적 수집
-- **현재 버전**: 4.2.6
+- **현재 버전**: 4.2.7
 
 ## 진행 현황 (2025-06-04)
 
@@ -208,7 +208,7 @@
   - 저장 위치: data/news_data/
   - 3가지 병합 모드 지원: sequential, even_distribution, recent_first
 
-### 🔄 현재 상태: Phase 1 코드 정리 거의 완료 (v4.2.6)
+### 🔄 현재 상태: Phase 1 코드 정리 완료 (v4.2.7)
 
 **✅ 주요 성과:**
 - 설정 파일 통합 완료 (3개 → 1개)
@@ -216,18 +216,12 @@
 - 문서 일관성 확보 완료
 - 코드 복잡도 감소 (65줄 순감소)
 
-**🎯 DEVELOPMENT_ROADMAP Phase 1 진행률: 85% 완료**
+**🎯 DEVELOPMENT_ROADMAP Phase 1 진행률: 100% 완료**
 
 프로젝트가 안정화 단계에 진입했습니다. 설정 관리 시스템이 단순화되어 유지보수성이 크게 향상되었습니다.
 
-#### Phase 1 남은 작업 (완료도 15%)
-1. **print문 → logging 통일** (우선순위: 🟡 중간)
-   - main.py, CLI 일부에서 여전히 print 사용
-   - 로깅 레벨 통일 및 일관성 확보
-   
-2. **테스트 코드 정리** (우선순위: 🟢 낮음)
-   - 루트 디렉토리의 test_*.py 파일들을 tests/ 디렉토리로 이동
-   - 테스트 코드 구조화 및 중복 제거
+#### Phase 1 남은 작업 (완료도 0%)
+모든 작업이 완료되었습니다.
 
 #### 다음 우선순위 작업 (Phase 1 완료 후)
 1. **Phase 2 시작 준비**: 미완성 기능 점검 및 우선순위 설정
@@ -320,4 +314,4 @@
 
 ---
 
-*최종 업데이트: 2025-06-04 (설정 파일 통합 완료 - v4.2.6)*
+*최종 업데이트: 2025-06-04 (Phase 1 코드 정리 완료 - v4.2.7)*
